+++ src/gui/ThemNhanVienDialog.java
@@ -129,7 +129,7 @@ public class ThemNhanVienDialog extends JDialog {
             }
             String gioiTinh = (String) cmbGioiTinh.getSelectedItem();
             String sdt = txtSdt.getText().trim();
-            float luong = Float.parseFloat(txtLuong.getText().trim());
+            double luong = Double.parseDouble(txtLuong.getText().trim());
             String diaChi = txtDiaChi.getText().trim();
             VaiTro vaiTro = (VaiTro) cmbVaiTro.getSelectedItem();
